Trying to add incremental loading (broken hop card click)
