+++ DruidsCornerApp/ViewModels/MainContext/ReferencesPageViewModel.cs
@@ -10,103 +10,23 @@ namespace DruidsCornerApp.ViewModels.MainContext;
 
 public partial class ReferencesPageViewModel : ObservableObject
 {
-    private readonly ILogger<RecipeExplorerViewModel> _logger;
-    private readonly HopProvider _hopProvider;
+    private readonly ILogger<ReferencesPageViewModel> _logger;
 
-    [ObservableProperty]
-    private ObservableCollection<CompactHopModel> _hops = new();
+    public HopReferenceViewModel HopReferenceViewModel { get; }
 
-    [ObservableProperty]
-    private ObservableCollection<string> _hopsNames = new();
-    
-    [ObservableProperty]
-    private HopReferenceFilters _hopFilters = new HopReferenceFilters();
-    
     [ObservableProperty]
     private int _selectedViewModelIndex = 0;
-
+    
     /// <summary>
     /// Base reference page (where the whole view is tabbed between hops, yeasts, malts and styles pages)
     /// </summary>
     /// <param name="logger"></param>
-    /// <param name="hopProvider"></param>
-    public ReferencesPageViewModel(ILogger<RecipeExplorerViewModel> logger,
-                                   HopProvider hopProvider
+    /// <param name="hopReferenceViewModel"></param>
+    public ReferencesPageViewModel(ILogger<ReferencesPageViewModel> logger,
+                                   HopReferenceViewModel hopReferenceViewModel
     )
     {
         _logger = logger;
-        _hopProvider = hopProvider;
-        InitFakeHops();
-    }
-    
-    private void InitFakeHops()
-    {
-        Hops.Clear();
-        var hops = _hopProvider.GetAllHops();
-
-        foreach (var hop in hops)
-        {
-            Hops.Add(CompactHopModelHelper.FromFullModel(hop));
-        }
-    }
-
-    [RelayCommand]
-    public Task RefreshData(CancellationToken cancellationToken)
-    {
-        InitFakeHops();
-        return Task.CompletedTask;
-    }
-
-    [RelayCommand]
-    public Task AddOneHop(CancellationToken cancellationToken)
-    {
-        Hops.Add(new CompactHopModel()
-        {
-            Name = "Fake hop!",
-            Purpose = "Don't know",
-            Rating = 4.1,
-            AlphaAcids = "9 - 12 %",
-            Favorite = true,
-            StockedAmount = 0
-        });
-        return Task.CompletedTask;
-    }
-
-    [RelayCommand]
-    public Task AddHopNameFilterLabel(Entry entryView, CancellationToken cancellationToken)
-    {
-        if (!HopsNames.Contains(entryView.Text) && !string.IsNullOrEmpty(entryView.Text))
-        {
-            HopsNames.Add(entryView.Text);
-            HopFilters.Names.Add(entryView.Text);
-        }
-
-        entryView.Text = "";
-        return Task.CompletedTask;
-    }
-
-    [RelayCommand]
-    public Task RemoveHopNameFilterLabel(string hopName)
-    {
-        if (HopsNames.Contains(hopName))
-        {
-            HopsNames.Remove(hopName);
-            HopFilters.Names.Remove(hopName);
-        }
-        return Task.CompletedTask;
-    }
-
-    [RelayCommand]
-    public async Task HopCardClicked(CompactHopModel hopModel)
-    {
-        await Shell.Current.GoToAsync($"HopPage?id={hopModel.Id}");
-    }
-
-    [RelayCommand]
-    public Task HopToggleFavorite(CompactHopModel hopModel)
-    {
-        hopModel.Favorite = !hopModel.Favorite;
-        
-        return Task.CompletedTask;
+        HopReferenceViewModel = hopReferenceViewModel;
     }
 }
