+++ DruidsCornerApp/Views/References/HopPage.xaml
@@ -102,7 +102,8 @@
                 <!-- Hop content -->
                 <Grid Grid.Row="1"
                       RowDefinitions="auto, auto, auto, auto, auto"
-                      RowSpacing="20">
+                      RowSpacing="20"
+                      Padding="10,0">
                     <!-- Hop attributes -->
                     <Grid Grid.Row="0"
                           RowDefinitions="auto, auto, auto, auto"
@@ -167,7 +168,7 @@
                         <Label Text="Tags"
                                FontSize="20"
                                FontAttributes="Bold"/>
-                        <CollectionView Grid.Row="1" 
+                        <CollectionView Grid.Row="1"
                                              ItemsSource="{Binding HopModel.Tags}">
                             <CollectionView.ItemsLayout>
                                 <LinearItemsLayout Orientation="Horizontal"
